requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
textract==1.6.5
